+++ plugins/nf-schema/src/main/nextflow/validation/Utils.groovy
@@ -270,6 +270,16 @@ public class Utils {
         return colorcodes
     }
 
+    public static String removeColors(String input) {
+        if (!input) {return input}
+        String output = input
+        List colors = logColours(false).collect { it.value }
+        colors.each { color ->
+            output = output.replace(color, "")
+        }
+        return output
+    }
+
     //
     // This function tries to read a JSON params file
     //